requirements.txt
numpy<2
lapx>=0.5.2
opencv-python==4.8.1.78
ultralytics==8.0.214
python-telegram-bot==20.3
python-dotenv==1.0.0
Pillow==10.1.0
pytest==7.4.3
